Convert ForinthryBracelet to kt plugin in order to be able to use attribute
